Remove ThrowIfNull for consistency with the rest of the project

+++ EngineBay.Blueprints.Tests/Bases/BaseTestWithDbContext.cs
@@ -15,8 +15,10 @@
                     .EnableSensitiveDataLogging()
                     .Options;
 
-            var context = Activator.CreateInstance(typeof(TContext), dbContextOptions) as TContext;
-            ArgumentNullException.ThrowIfNull(context);
+            if (Activator.CreateInstance(typeof(TContext), dbContextOptions) is not TContext context)
+            {
+                throw new ArgumentException("Context is null");
+            }
 
             this.DbContext = context;
             this.DbContext.Database.EnsureDeleted();
